Create csv_reader.py
Hi Mohammad! How are you? Here is my code for the csv reader. Not the final code version, so I will continue editing it today and tomorrow. I just have a question - here, I assumed that "cifar10_experiment_results.csv" and "imagenet_experiment_results.csv" were the csvs with pairs that we needed to use, but I just want to double check with you if they are the correct csv files. 
Thank you for your time and have a wonderful day!