Update csv_reader.py
Hello Mohammad! I hope this message finds you well.
Here is an updated version of the csv_reader.py:

1) Seems to work fine now as it correctly pairs images together. 
2) It can now download all the images on the csv file unless the csv_images folder exists. 
3) Instead of the PIL image or the url, the relative paths for img1 and img2 are stored in the dictionaries. 

However, I still had a few issues with Matplotlib since it wouldn't visualize the images and their label. I did some debugging and it seems like this an environment issue on my end. Overall, it should work fine since it prints the correct paths to the images.