Update metrics.py
I think this might solve the issue for ssim. I read that the reason why it was giving errors was because ssim wanted img_as_float of grayscale images (only 1 channel), not the coloured ones (3 channels for example). There is also an alternative for coloured images added as well, but it's commented out.
All the other metrics that were commented out seem fine.